pyproject.toml
[project]
name = "key"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.128.1",
    "opencv-python>=4.13.0.92",
    "openfoodfacts>=4.0.0",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.24",
    "pyzbar>=0.1.9",
    "sentence-transformers>=5.3.0",
    "zai-sdk>=0.2.2",
]
